+++ fsl/utils/image/__init__.py
@@ -1,0 +1,16 @@
+#!/usr/bin/env python
+#
+# __init__.py - The fsl.utils.image package
+#
+# Author: Paul McCarthy <pauldmccarthy@gmail.com>
+#
+"""The :mod:`fsl.utils.image` oackage contains algorithms and utilities for
+manipulating and working with :class:`.Image` objects.
+
+The following modules are available:
+
+.. autosumary::
+   :nosignature
+
+   .image.resample
+"""