+++ CHANGELOG.rst
@@ -2,6 +2,49 @@ This document contains the ``fslpy`` release history in reverse chronological
 order.
 
 
+2.2.0 (Wednesday May 8th 2019)
+------------------------------
+
+
+Added
+^^^^^
+
+
+* New :mod:`.resample_image` script.
+* New :mod:`.resample` module (replacing the :func:`.Image.resample` method),
+  containing functions to resample an :class:`.Image`.
+* New :func:`.resample.resampleToPixdim` and
+  :func:`.resample.resampleToReference` functions, convenience wrappers around
+  :func:`.resample.resample`.
+* New :func:`.idle.block` function.
+
+
+Changed
+^^^^^^^
+
+
+* The :func:`.resample` function (formerly :meth:`.Image.resample`) now
+  accepts ``origin`` and ``matrix`` parameters, which can be used to adjust
+  the alignment of the voxel grids of the input and output images.
+* The :func:`.transform.decompose` function now accepts both ``(3, 3)``
+  and ``(4, 4)`` matrices.
+
+
+Fixed
+^^^^^
+
+
+* Minor fixes to some :mod:`.filetree` tree definitions.
+
+
+Deprecated
+^^^^^^^^^^
+
+
+* The :meth:`.Image.resample` method has been deprecated in favour of the
+  :func:`.resample.resample` function.
+
+
 2.1.0 (Saturday April 13th 2019)
 --------------------------------